pyproject.toml
[tool.black]
line-length = 120
target-version = ["py311"]
skip-string-normalization = true
exclude = '''
^/(
  (
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | alembic
    | backend_utils
    | venv
  )/
)
'''

[tool.isort]
line_length = 120
combine_as_imports = true
order_by_type = true
length_sort_straight = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
known_local_folder = ['webapp', 'backend_utils', 'conf']
extend_skip = ["backend_utils", "alembic", "*pypoetry*"]

[tool.poetry]
name = "sirius"
version = "0.1.0"
description = ""
authors = ["Gladkikh Anton"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.103.1"
uvicorn = "0.23.1"
uvloop = "0.17.0"
pydantic = { extras = ["dotenv"], version = "2.3.0" }
pydantic-settings = "2.0.3"
orjson = "3.9.7"
poetry = "1.5.1"
aiohttp = "3.8.5"
python-jose = "3.3.0"
redis = "5.0.1"
asyncpg = "^0.28.0"
alembic = "1.12.1"
sqlalchemy = "2.0.23"
aiokafka = "0.8.1"
python-multipart = "0.0.6"
starlette-prometheus = "0.9.0"
starlette-context = "0.3.6"
miniopy-async = "1.17"
minio = "^7.2.5"

[tool.poetry.group.dev.dependencies]
autoflake = "2.2.0"
black = "23.7.0"
flake8 = "6.0.0"
flake8-black = "0.3.6"
flake8-bugbear = "23.7.10"
flake8-coding = "1.3.2"
flake8-comprehensions = "3.14.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.5.0"
flake8-executable = "2.1.3"
flake8-isort = "6.0.0"
flake8-logging-format = "0.9.0"
flake8-pie = "0.16.0"
flake8-pytest-style = "1.7.2"
flake8-quotes = "3.3.2"
flake8-simplify = "0.20.0"
isort = "5.12.0"
mypy = "1.5.1"
pytest = "^7.1.2"
aio-pika = "9.4.0"
pytest-env = "^0.6.2"
pytest-cov = "3.0.0"
pytest-asyncio = "0.20.3"
pytest-freezegun = "0.4.2"
types-python-jose = "3.3.4.8"
psycopg2-binary = "^2.9.6"
httpx = "0.25.2"

[tool.pytest.ini_options]
addopts = "--failed-first --exitfirst --showlocals --cov=."
asyncio_mode = "auto"
python_files = "tests.py test_*.py"
python_functions = "test_*"
testpaths = ["tests"]
env = [

]

[tool.coverage.run]
omit = [
    #"alembic*/*",
    #"scripts/*",
    #"tests/*",
    #"webapp/__main__.py", # local run
]

[tool.mypy]
plugins = [
    "pydantic.mypy",
]
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_reexport = false
exclude = [
    # "^alembic/",
    # "^conf/",
    # "^tests/",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = false  # to allow parsing
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = [
    "aiokafka.*",
    "aiocache.*",
    "aioredis.*",
    "aioretry.*",
    "asyncpg.*", # https://github.com/MagicStack/asyncpg/issues/569
    "cache.*",
    "gunicorn.*",
    "msgpack",
    "prometheus_client.*",
    "pythonjsonlogger.*",
    "starlette_prometheus.*",
    "uvicorn.*",
    "xxhash.*",
]
ignore_missing_imports = true
